requirements.txt
pyrogram==2.0.49
pytgcalls==3.4.0
yt-dlp==2025.12.31
python-dotenv==1.0.0
ffmpeg-python==0.3.0